requirements.txt
# requirements.txt
ccxt       # Integração com exchanges (Bitget) [[7]]
pandas      # Manipulação de dados em tempo real
numpy     # Cálculos numéricos avançados
schedule     # Agendamento de operações
python-decouple # Gerenciamento de variáveis de ambiente [[5]]
loguru       # Logs estruturados
pytest       # Testes automatizados
ta          # Indicadores técnicos (RSI, MACD, etc.)
matplotlib 
pytest-mock # Permite usar mocker em testes
mkdocs
mkdocs-material 
pdoc
pdoc3
streamlit
streamlit-autorefresh
plotly
websocket-client
python-telegram-bot
asyncio
mplfinance
pytest-asyncio
pydantic
aiofiles